@@ -1,7 +1,7 @@
 {
   "name": "postgres-array",
   "main": "index.js",
-  "version": "2.0.0",
+  "version": "3.0.4",
   "description": "Parse postgres array columns",
   "license": "MIT",
   "repository": "bendrucker/postgres-array",
@@ -11,7 +11,7 @@
     "url": "bendrucker.me"
   },
   "engines": {
-    "node": ">=4"
+    "node": ">=12"
   },
   "scripts": {
     "test": "standard && tape test.js"
@@ -24,8 +24,8 @@
   ],
   "dependencies": {},
   "devDependencies": {
-    "standard": "^12.0.1",
-    "tape": "^4.0.0"
+    "standard": "^17.0.0",
+    "tape": "^5.0.0"
   },
   "files": [
     "index.js",
