@@ -1,7 +1,7 @@
 {
   "name": "postgres-bytea",
   "main": "index.js",
-  "version": "1.0.0",
+  "version": "3.0.0",
   "description": "Postgres bytea parser",
   "license": "MIT",
   "repository": "bendrucker/postgres-bytea",
@@ -11,10 +11,10 @@
     "url": "bendrucker.me"
   },
   "engines": {
-    "node": ">=0.10.0"
+    "node": ">= 6"
   },
   "scripts": {
-    "test": "standard && tape test.js"
+    "test": "standard && tape *.js"
   },
   "keywords": [
     "bytea",
@@ -22,13 +22,17 @@
     "binary",
     "parser"
   ],
-  "dependencies": {},
+  "dependencies": {
+    "obuf": "~1.1.2"
+  },
   "devDependencies": {
-    "tape": "^4.0.0",
-    "standard": "^4.0.0"
+    "concat-stream": "2.0.0",
+    "standard": "^14.0.0",
+    "stream-to-promise": "^3.0.0",
+    "tape": "^5.0.0",
+    "tape-promise": "4.0.0"
   },
   "files": [
-    "index.js",
-    "readme.md"
+    "*.js"
   ]
 }
