@@ -1,17 +1,17 @@
 {
   "name": "postgres-interval",
   "main": "index.js",
-  "version": "1.2.0",
+  "version": "3.0.0",
   "description": "Parse Postgres interval columns",
   "license": "MIT",
   "repository": "bendrucker/postgres-interval",
   "author": {
     "name": "Ben Drucker",
     "email": "bvdrucker@gmail.com",
-    "url": "bendrucker.me"
+    "url": "https://www.bendrucker.me"
   },
   "engines": {
-    "node": ">=0.10.0"
+    "node": ">=12"
   },
   "scripts": {
     "test": "standard && tape test.js"
@@ -21,12 +21,10 @@
     "interval",
     "parser"
   ],
-  "dependencies": {
-    "xtend": "^4.0.0"
-  },
+  "dependencies": {},
   "devDependencies": {
-    "tape": "^4.0.0",
-    "standard": "^12.0.1"
+    "standard": "^16.0.0",
+    "tape": "^5.0.0"
   },
   "files": [
     "index.js",
